Manual add of changes in commit 0c46e9dac1

+++ src/generate_report.pm
@@ -6,17 +6,21 @@
     my $lsf_out=$lsf_file_dir."/".$current_job_file.".out";
     my $lsf_err=$lsf_file_dir."/".$current_job_file.".err";
     `rm $lsf_out`;
+    if(-e $lsf_out) 
+	{
+	`rm $lsf_out`;
     `rm $lsf_err`;
     `rm $current_job_file`;
+	}
     open(REPRUN, ">$job_files_dir/$current_job_file") or die $!;
     print REPRUN "#!/bin/bash\n";
     print REPRUN "#BSUB -n 1\n";
     print REPRUN "#BSUB -R \"rusage[mem=40000]\"","\n";
     print REPRUN "#BSUB -M 40000000\n";
-    print REPRUN "#BSUB -a \'docker(registry.gsc.wustl.edu/genome/genome_perl_environment)\'\n";
+    #print REPRUN "#BSUB -a \'docker(registry.gsc.wustl.edu/genome/genome_perl_environment)\'\n";
     #print STREKA "#BSUB -q long\n";
-    print REPRUN "#BSUB -q research-hpc\n";
-    #print REPRUN "#BSUB -q ding-lab\n";
+    #print REPRUN "#BSUB -q research-hpc\n";
+    print REPRUN "#BSUB -q ding-lab\n";
     print REPRUN "#BSUB -o $lsf_file_dir","/","$current_job_file.out\n";
     print REPRUN "#BSUB -e $lsf_file_dir","/","$current_job_file.err\n";
     print REPRUN "#BSUB -J $current_job_file\n";
